relocated util functions to utils.py
